Add Arr
Нужна помощь в преобразовании переменной в строку, также нужна проверка.

+++ task/task/Device.cs
@@ -12,7 +12,7 @@ namespace task
 
         string Type { get; set; }
 
-        double Price { get; set; }
+        float Price { get; set; }
 
     }
